add prop types here and add random ID for key
